Refactoring - move to service layer
