+++ openpkw-services/src/main/java/org/openpkw/qr/service/QrService.java
@@ -1,0 +1,11 @@
+package org.openpkw.qr.service;
+
+/**
+ * Service for parsing and saving qr from mobile
+ * @author Sebastian Pogorzelski
+ */
+public interface QrService {
+
+
+
+}
